LayoutTests/accessibility/input-type-hidden-in-aria-hidden-false.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/accessibility-helper.js"></script>
<script src="../resources/js-test.js"></script>
</head>
<body>

<div id="content" role="group">
   <input type="hidden">
</div>

<script>
    var testOutput = "This tests that input type of hidden are not exposed when aria-hidden=false\n\n";

    if (window.accessibilityController) {
        window.jsTestIsAsync = true;

        // By default, this should have no children because input type=hidden is hidden.
        var content = accessibilityController.accessibleElementById("content");
        testOutput += expect("content.childrenCount", "0");

        // When aria-hidden=true, content should not even be there
        document.getElementById("content").setAttribute("aria-hidden", "true");
        setTimeout(async function() {
            await waitFor(() => {
               content = accessibilityController.accessibleElementById("content");
               return !content || !content.isValid();
            });
            testOutput += expect("!content || !content.isValid()", "true");

            // When aria-hidden=false, we should NOT expose input type=hidden
            document.getElementById("content").setAttribute("aria-hidden", "false");
            await waitFor(() => {
               content = accessibilityController.accessibleElementById("content");
               return content.childrenCount === 0;
            });
            testOutput += expect("content.childrenCount", "0");

            // And changing back to true on the fly should have a similar effect
            document.getElementById("content").setAttribute("aria-hidden", "true");
            await waitFor(() => {
               content = accessibilityController.accessibleElementById("content");
               return !content || !content.isValid();
            });
            testOutput += expect("!content || !content.isValid()", "true");

            debug(testOutput);
            finishJSTest();
        }, 0);
    }
</script>
</body>
</html>
