Safari-609.1.11.1/LayoutTests/accessibility/input-type-hidden-in-aria-hidden-false-expected.txt

This tests that input type of hidden are not exposed when aria-hidden=false

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS content.childrenCount is 0
PASS !content || !content.isValid() is true
PASS content.childrenCount is 0
PASS !content || !content.isValid() is true
PASS successfullyParsed is true

TEST COMPLETE
